// src/pages/ArticleManagement/ArticleList/data.d.ts
export interface dataType {

  /**
   * 添加时间
   */
  addTime: string;

  /**
   * 文章id
   */
  id: number;
  /**
   * 文章简介
   */
  introduce: string;

  /**
   * 文章标题
   */
  title: string;

  /**
   * 文章类型
   */
  typeName: string;

  /**
   * 文章阅读量
   */
  view_count: number;

  /**
   * 文章状态
   */
  status: number;
}